pyproject.toml
[project]
name = "src"
version = "0.0.0"
description = "Code of the paper 'Estimating Individual Tree Height and Species from UAV Imagery'"
authors = [{ name = "Jannik Endres", email = "jannik.endres@mila.quebec" }]
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = "Apache-2.0"
dependencies = [
    "albumentations==2.0.8",
    "hydra-core==1.3.2",
    "lightning==2.5.2",
    "matplotlib==3.10.3",
    "numpy==2.0.1",
    "omegaconf==2.3.0",
    "pandas==2.3.1",
    "pre-commit>=4.2.0,<5",
    "rasterio==1.4.2",
    "scikit-learn==1.7.1",
    "shapely==2.1.1",
    "timm==1.0.19",
    "torch==2.8.0",
    "torchmetrics==1.8.1",
    "torchvision==0.23.0",
    "wandb==0.21.1",
]

[dependency-groups]
dev = []

[project.optional-dependencies]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


#
# Scripts
#
[project.scripts]


#
# External
#
[tool.uv.workspace]
members = [
    "external/facebookresearch_dinov3",
    "external/hugobaudchon_geodataset",
]

#
# Tools
#

[tool.autoflake]
remove-all-unused-imports = true
in-place = true
ignore-init-module-imports = true
remove-unused-variables = true
recursive = true

[tool.autopep8]
max_line_length = 120
in-place = true
aggressive = 2
ignore = ["W503", "E203", "E501"]
recursive = true

[tool.flake8]
max-line-length = 120
ignore = ["E203", "E266", "E501", "W503"]
docstring-convention = "numpy"
max-complexity = 20
per-file-ignores = []
exclude = [
    ".git",
    "migrations",
    "__pycache__"
]

[tool.pylint]
disable = "C0114,C0115,C0116,R0903,R1710,W1203"
max-line-length = 120
max-locals = 20
max-args = 16
max-positional-arguments = 16
max-attributes = 20

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"

[tool.flynt]
line-length = 120
transform-concats = true
verbose = true

[tool.docformatter]
style = "numpy"
pre-summary-newline = true
wrap-descriptions = 120
wrap-summaries = 120
blank = false
exclude = []
recursive = true